refactor(ARCO-282): Separate callbacker #642
